services/model-server/requirements.txt
# fastapi==0.104.1
# uvicorn==0.24.0
# pydantic==2.5.0
# numpy==1.24.3
# scikit-learn==1.3.0
# joblib==1.3.2
# redis==5.0.1
# requests==2.31.0

fastapi 
uvicorn 
scikit-learn 
redis 
numpy 
joblib 
pydantic
requests
flair


boto3
python-dotenv